Remove codersrank section
Maintenance of the service is not sufficient, widget have been broken
for some time.  This was nice when it worked, but it is now clueless, so
I removed my account.

OpenHub has relevant data but no practical way to integrate it here, so
for now just do not include such stats.